QR code  (#6)
* add packages

* add packages

* add course and certificate packages

* add course and certificate packages

* fix certificate packages

* fix course packages

* add resources packages

* add phase packages

* add place packages

* add question packages

* add quiz packages

* add answer packages

* fix certificate crud

* fix crud answer,phase,course,place,resources,question

* add courseCategory crud

* add courseProgress crud

* add courseComment crud

* fix bugs

* fix bugs

* fix bugs

* fix crud phase

* fix answer crud

* fix ressources crud

* fix quiz crud

* fix crud

* fix crud

* add QR code

+++ src/main/java/tn/dalhia/services/implementations/PhaseService.java
@@ -3,7 +3,9 @@ package tn.dalhia.services.implementations;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import tn.dalhia.entities.Course;
 import tn.dalhia.entities.Phase;
+import tn.dalhia.repositories.CourseRepository;
 import tn.dalhia.repositories.PhaseRepository;
 import tn.dalhia.services.IPhaseService;
 
@@ -14,6 +16,8 @@ public class PhaseService implements IPhaseService {
 
     @Autowired
     PhaseRepository phaseRepository;
+    @Autowired
+    CourseRepository courseRepository;
 
     @Override
     public List<Phase> getAll(){
@@ -21,9 +25,16 @@ public class PhaseService implements IPhaseService {
     }
 
     @Override
-    public Phase add(Phase phase){
+    public Phase add(Phase phase, Long id){
+        Course c = courseRepository.findById(id).orElse(null);
+        if (c == null)
+            return null;
 
-        return phaseRepository.save(phase);
+        phase.setNumber(c.getNbrPhases()+1);
+        c.getPhases().add(phase);
+        c.setNbrPhases(c.getNbrPhases()+1);
+        courseRepository.save(c);
+        return phase;
     }
 
     @Override
@@ -41,9 +52,22 @@ public class PhaseService implements IPhaseService {
     }
 
     @Override
-    public  boolean delete(Long id){
+    public List<Phase> getAllByCourse(Long id){
+        Course c = courseRepository.findById(id).orElse(null);
+        if (c == null)
+            return null;
+
+        return c.getPhases();
+    }
+
+    @Override
+    public  boolean delete(Long id, Long idCourse){
         Phase ph = phaseRepository.findById(id).orElse(null);
-        if(ph != null){
+        Course c = courseRepository.findById(idCourse).orElse(null);
+
+        if(ph != null && c!=null){
+            c.setNbrPhases(c.getNbrPhases()-1);
+            courseRepository.save(c);
             phaseRepository.delete(ph);
             return true;
         }
